New Method for Scrolling
Scrolling is done by drawing the map in its full size, drawing everything
onto it, and then blitting only the portion of it that needs to be seen onto
the screen. (For a detailed explanation see
https://github.com/Inglorion35/scrolling-demo).

These changes still need to be merged with ones that were made while
this method was being implemented; hence the creation of a seperate
branch for it.