Perfected camera, moving cloud
The camera now adjusts its position so that it does not display outside
the map when you are at the edges, and centers on the player under
all other circumstances. Also, one cloud with a random velocity is
created at the starting; more work is needed to have multiple,
randomized spawns.

Work has started on decreasing air density at higher altitudes;
those lines are commented out.