Fixed terrain collision; constants module; birds
Commonly used variables are initialized in
constants.py which is referenced in
game_sprites.py, phy.py, gamemenu.py and init.py
for better code. constants.py is imported in
those files.

The weird issues surrounding terrain collision
are now fixed; it now seems to be working exactly
as expected and no further work is needed.

The code for birds has been added; it is
commented out as it has as-yet unknown issues
that appear when the Bird class's update
function is called.

Reverting the commit means the clouds function
needs to be re-implemented again with a proper
random number generation, which can be reused
for birds as well. Must not spawn very close to
player and spawn rate needs to be fixed.