Finished cloud RNG; Implemented Birds, misc.
Cloud RNG ensures that clouds spawn outside
the view of the player, and prevents them
from spawning inside terrain by doing a
one-time collision detection with the terrain
mask.

Trial: A for-loop to import a collection of
cloud sprites to randomize the image for
each cloud that spawns

We've gotten a single bird sprite to spawn and
move in a zig-zag motion. It's collision also
works; RNG spawning has to be implemented next.

The game background is now loaded; however it
has to be loaded in low resolution, otherwise
memory limit is exceeded. Further work is needed.

Decelerating has been fixed in game_sprites.
The event handling block has been altered.

TODO:
check notes.txt