Code restructuring/cleanup
Commit 5507039 had moved the loading of cloud
and bird sprites to game_sprites so that they
could be used in cloudspawn and birdspawn;
loading has now been moved back to init and
the sprites are supplied as arguments to these
functions when they are called in init.

Also, the redundant 'self' parameter has been
removed from those two functions.