Optimize terrain blitting (fps increase)
Now, the game only blits the portion of the terrain
that is supposed to be visible (as defined by the
camera rect) to the screen, rather than the whole
thing. This drastically improved FPS from ~30 to
~ 60.
However, this also introduces a minor bug where
you can see a tiny black portion at the screen edges
when you scroll to them; probably a result of the
camera rect not updating in time. Fixing this is
not high priority.

See notes.txt for current tasks

Updated notes.txt with current